add query for aborted amount in zeta
