+++ proto/crosschain/query.proto
@@ -147,6 +147,19 @@ service Query {
   rpc CctxByStatus(QueryCctxByStatusRequest) returns (QueryCctxByStatusResponse) {
     option (google.api.http).get = "/zeta-chain/crosschain/cctxbyStatus/{status}";
   }
+
+  rpc AbortedZetaAmount(QueryAbortedZetaAmountRequest) returns (QueryAbortedZetaAmountResponse) {
+    option (google.api.http).get = "/zeta-chain/crosschain/abortedZetaAmount";
+  }
+}
+
+message QueryAbortedZetaAmountRequest {}
+
+message QueryAbortedZetaAmountResponse {
+  string amount = 6 [
+    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Uint",
+    (gogoproto.nullable) = false
+  ];
 }
 
 message QueryCctxByStatusRequest {
