+++ zetaclient/zeta_supply_checker.go
@@ -175,17 +175,13 @@ func ValidateZetaSupply(logger zerolog.Logger, abortedTxAmounts, zetaInTransit, 
 }
 
 func (zs *ZetaSupplyChecker) AbortedTxAmount() (sdkmath.Int, error) {
-	cctxList, err := zs.zetaClient.GetCctxByStatus(types.CctxStatus_Aborted)
+	amount, err := zs.zetaClient.GetAbortedZetaAmount()
 	if err != nil {
-		return sdkmath.ZeroInt(), err
-	}
-	amount := sdkmath.ZeroUint()
-	for _, cctx := range cctxList {
-		amount = amount.Add(cctx.GetCurrentOutTxParam().Amount)
+		return sdkmath.ZeroInt(), errors.Wrap(err, "error getting aborted zeta amount")
 	}
 	amountInt, ok := sdkmath.NewIntFromString(amount.String())
 	if !ok {
-		return sdkmath.ZeroInt(), errors.New("error parsing amount")
+		return sdkmath.ZeroInt(), errors.New("error parsing aborted zeta amount")
 	}
 	return amountInt, nil
 }
